separate utils to avoid dep cycles
