+++ src/containers/game/utils/index.ts
@@ -1,0 +1,3 @@
+export { default as calculateGameScale } from './calculateGameScale';
+export { default as getCharacterPropsFromKeyCode } from './getCharacterPropsFromKeyCode';
+export { default as getMapConfig } from './getMapConfig';
